refactor: move connector search to dedicated module
