+++ turbo/packages/core/src/index.ts
@@ -236,7 +236,6 @@ export {
   getScopeDiff,
   getApiTokenFieldsByType,
   deriveApiTokenConnectedTypes,
-  searchConnectors,
   connectorSessionStatusSchema,
   connectorResponseSchema,
   connectorListResponseSchema,
@@ -579,8 +578,6 @@ export {
   type ConnectorAuthMethodConfig,
   type ConnectorAuthMethodId,
   type ScopeDiff,
-  type ConnectorSearchResult,
-  type ConnectorSearchOutput,
   type ScopeDiffResponse,
   type ConnectorSessionStatus,
   type ConnectorResponse,
